added home component
